delivery_service/requirements.txt
psycopg2-binary==2.9.9
django==5.0.3
djangorestframework==3.15.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0